Update gson, move source to src/main/java directory
