+++ .gitignore
@@ -57,3 +57,6 @@ local.properties
 .cache-main
 .scala_dependencies
 .worksheet
+
+# IntelliJ IDEA
+.idea/
