+++ pom.xml
@@ -5,7 +5,6 @@
   <version>1.5</version>
   <name>Zermelo4J</name>
   <build>
-    <sourceDirectory>src</sourceDirectory>
     <plugins>
       <plugin>
         <artifactId>maven-compiler-plugin</artifactId>
@@ -21,7 +20,7 @@
 	<dependency>
 	  <groupId>com.google.code.gson</groupId>
 	  <artifactId>gson</artifactId>
-	  <version>2.8.5</version>
+	  <version>2.9.0</version>
 	</dependency>
   </dependencies>
 </project>
