+++ src/main/java/nl/mrwouter/zermelo4j/annoucements/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Package that contains the announcement object for Zermelo4J
+ */
+package nl.mrwouter.zermelo4j.annoucements;
