+++ src/main/java/nl/mrwouter/zermelo4j/appointments/AppointmentParticipationException.java
