+++ src/main/java/nl/mrwouter/zermelo4j/appointments/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Package that contains the appointment objects for Zermelo4J
+ */
+package nl.mrwouter.zermelo4j.appointments;
