+++ src/main/java/nl/mrwouter/zermelo4j/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Main package for the Zermelo4J project
+ */
+package nl.mrwouter.zermelo4j;
