--- src/nl/mrwouter/zermelo4j/appointments/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Package that contains the appointment objects for Zermelo4J
- */
-package nl.mrwouter.zermelo4j.appointments;
