--- src/nl/mrwouter/zermelo4j/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Main package for the Zermelo4J project
- */
-package nl.mrwouter.zermelo4j;
